% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{Acells}
\alias{Acells}
\title{Identify non-NA values in a raster}
\usage{
Acells(object)
}
\arguments{
\item{object}{raster object.}
}
\value{
vector of cells which values aren't NA.
}
\description{
\code{Acells} returns the cells that don't have NA values in a raster object.
}
