% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{addGenealSimProb}
\alias{addGenealSimProb}
\alias{addGenealSimProb,ecogenetSetLik,genealSimProb}
\title{addGenealSimProb method for ecogenetSetLik objects.}
\usage{
\S4method{addGenealSimProb}{ecogenetSetLik,genealSimProb}(object, genSimProb)
}
\description{
addGenealSimProb method for ecogenetSetLik objects.
}
