% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{addGenealSimProb}
\alias{addGenealSimProb}
\title{Adds genealSimProb object to list}
\usage{
addGenealSimProb(object, genSimProb)
}
\arguments{
\item{object}{list to append the genealSimProb object.}

\item{genSimProb}{genealSimProb object to append.}
}
\value{
list plus the appended object.
}
\description{
This method adds a genealSimProb object to a list.
}
