% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{addSimulCoal}
\alias{addSimulCoal}
\alias{addSimulCoal,genealSimProb,logical,numeric,logical}
\alias{addSimulCoal,genealSimProb,logical,numeric}
\alias{addSimulCoal,ecogenetSetLik,logical,numeric,ANY}
\alias{addSimulCoal,ecogenetSetLik,logical,numeric}
\title{addSimulCoal method for genealSimProb object.}
\usage{
\S4method{addSimulCoal}{genealSimProb,logical,numeric,missing,ANY}(object, printCoal, iter, multiCore = FALSE, cores = NULL)

\S4method{addSimulCoal}{genealSimProb,logical,numeric,logical,missing}(object, printCoal, iter, multiCore = FALSE, cores = NULL)

\S4method{addSimulCoal}{genealSimProb,logical,numeric,logical,numeric}(object, printCoal, iter, multiCore = FALSE, cores = NULL)

\S4method{addSimulCoal}{ecogenetSetLik,logical,numeric,missing,ANY}(object, printCoal, iter, multiCore = FALSE, cores = NULL)

\S4method{addSimulCoal}{ecogenetSetLik,logical,numeric,logical,missing}(object, printCoal, iter, multiCore = FALSE, cores = NULL)

\S4method{addSimulCoal}{ecogenetSetLik,logical,numeric,logical,numeric}(object, printCoal, iter, multiCore = FALSE, cores = NULL)
}
\arguments{
\item{object}{ecogenetSetLik object.}
}
\value{
genealSimProb object with the added coalescent simulations.

genealSimProb object with the added coalescent simulations.

genealSimProb object with the added coalescent simulations.

ecogenetSetLik object with the added coalescent simulations.

ecogenetSetLik object with the added coalescent simulations.

ecogenetSetLik object with the added coalescent simulations.
}
\description{
addSimulCoal method for genealSimProb object.

addSimulCoal method for genealSimProb object.

addSimulCoal method for genealSimProb object.

addSimulCoal method for ecogenetSetLik object.

addSimulCoal method for ecogenetSetLik object.

addSimulCoal method for ecogenetSetLik object.
}
