% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{addSimulCoal}
\alias{addSimulCoal}
\title{Adds coalescent simulations to an object}
\usage{
addSimulCoal(object, printCoal, iter, multiCore = FALSE, cores = NULL)
}
\arguments{
\item{object}{the object to add the coalescent simulations.}

\item{printCoal}{boolean. Whether or not print the coalescing times during the simulation.}

\item{iter}{numeric. Amount of simulations to perform.}

\item{multiCore}{boolean. Whether to use multicore execution. This cannot be done in Windows. Defaults to \code{FALSE}.}

\item{cores}{numeric. Amount of cores to use for the simulations.}
}
\description{
This function simulates multiple coalescents, calculates their corresponding log-likelihood and adds them to an object.
}
