% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{buildGeodist}
\alias{buildGeodist}
\alias{buildGeodist,socioecoGeoDataModel}
\title{buildGeodist method for socioecoGeoDataModel objects.}
\usage{
\S4method{buildGeodist}{socioecoGeoDataModel}(object)
}
\description{
buildGeodist method for socioecoGeoDataModel objects.
}
