% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{buildGeodist}
\alias{buildGeodist}
\title{Build geographic distance matrix}
\usage{
buildGeodist(object)
}
\arguments{
\item{object}{Object containing raster data to calculate the geographical distance.}
}
\value{
Matrix of dimension n^2 (being n the number of cells in the raster) with the geographic distance among cells.
}
\description{
Builds a geographic distance matrix from raster data.
}
