% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{buildMigrationMatrix}
\alias{buildMigrationMatrix}
\alias{buildMigrationMatrix,socioecoGeoDataModel}
\title{buildMigrationMatrix method for socioecoGeoDataModel objects.}
\usage{
\S4method{buildMigrationMatrix}{socioecoGeoDataModel}(object)
}
\description{
buildMigrationMatrix method for socioecoGeoDataModel objects.
}
