% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{buildMigrationMatrix}
\alias{buildMigrationMatrix}
\title{Build Migration Matrix}
\usage{
buildMigrationMatrix(object)
}
\arguments{
\item{object}{Object containing raster data to build the migration probability matrix,}
}
\value{
Matrix of size n^2 * n^2 (being n the number of cells) containing the transition probability among cells.
}
\description{
Build a migration probability matrix for the landscape described in a raster.
}
