% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{buildRKlandscape}
\alias{buildRKlandscape}
\alias{buildRKlandscape,socioecoGeoDataModel}
\title{buildRKlandscape method for socioecoGeoDataModel objects.}
\usage{
\S4method{buildRKlandscape}{socioecoGeoDataModel}(object)
}
\description{
buildRKlandscape method for socioecoGeoDataModel objects.
}
