% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{buildRKlandscape}
\alias{buildRKlandscape}
\title{Build an R/K landscape}
\usage{
buildRKlandscape(object)
}
\arguments{
\item{object}{Object containing the information required to build the R/K landscape.}
}
\value{
rasterStack object containing the R lanscape in a layer and the K landscape in another layer.
}
\description{
The niche function makes the product of the different layers results
}
