% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{buildTransitionBackward}
\alias{buildTransitionBackward}
\alias{buildTransitionBackward,socioecoGeoDataModel}
\title{buildTransitionBackward method for socioecoGeoDataModel objects.}
\usage{
\S4method{buildTransitionBackward}{socioecoGeoDataModel}(object)
}
\description{
buildTransitionBackward method for socioecoGeoDataModel objects.
}
