% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{buildTransitionBackward}
\alias{buildTransitionBackward}
\title{Build a backwards transition matrix}
\usage{
buildTransitionBackward(object)
}
\arguments{
\item{object}{Object containing the landscape and niche information.}
}
\value{
Backwards transition matrix.
}
\description{
Builds a backwards transition matrix from landscape and niche information.
}
