% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{buildTransitionForward}
\alias{buildTransitionForward}
\alias{buildTransitionForward,socioecoGeoDataModel,character}
\title{buildTransitionForward method for socioecoGeoDataModel objects.}
\usage{
\S4method{buildTransitionForward}{socioecoGeoDataModel,character}(object, meth)
}
\description{
buildTransitionForward method for socioecoGeoDataModel objects.
}
