% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{buildTransitionForward}
\alias{buildTransitionForward}
\title{Build a forward transition matrix}
\usage{
buildTransitionForward(object, meth)
}
\arguments{
\item{object}{Object containing the landscape and niche information.}
}
\value{
Forward transition matrix.
}
\description{
Builds a forward transition matrix from landscape and niche information.
}
