% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{categories}
\alias{categories}
\alias{categories,socioecoGroupData}
\alias{categories,socioecoGroupsData}
\alias{categories,geoEnvData}
\title{Categories method for socioecoGroupData.}
\usage{
\S4method{categories}{socioecoGroupData}(object)

\S4method{categories}{socioecoGroupsData}(object)

\S4method{categories}{geoEnvData}(object)
}
\description{
Categories method for socioecoGroupData.

Categories method for socioecoGroupsData.

Categories method for geoEnvData objects.
}
