% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{categories}
\alias{categories}
\title{Method to obtain the category names of the connection types.}
\usage{
categories(object)
}
\arguments{
\item{object}{Any object with a \code{categories} slot.}
}
\value{
names of the values of the \code{categories} slot.
}
\description{
Method to obtain the category names of the non-geographic connection types.
}
