% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{coalescent_2_newick}
\alias{coalescent_2_newick}
\alias{coalescent_2_newick,coalescent}
\title{coalescent_2_newick method for coalescent objects.}
\usage{
\S4method{coalescent_2_newick}{coalescent}(object)
}
\description{
coalescent_2_newick method for coalescent objects.
}
