% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{coalescent_2_newick}
\alias{coalescent_2_newick}
\title{Converts coalSim objects into Newick trees.}
\usage{
coalescent_2_newick(object)
}
\arguments{
\item{coalescent}{coalSim object. This object is generated by a coalescent simulation (implemented in the \code{simulCoal} method).}
}
\value{
Newick format tree.
}
\description{
This method converts coalSim objects from a coalescence simulation into a Newick-formatted tree.
}
