% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{coalSim-class}
\alias{coalSim-class}
\title{coalSim class}
\description{
This class is a container for the coalescence simulations.
}
\section{Slots}{

\describe{
\item{\code{coalescent}}{list. This list stores the coalescent events and times from a coalescent calculated by the \code{simulCoal} method.}

\item{\code{probForward}}{numeric. Forward demographic probability of the coalescent simulation.}
}}

