% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{ecoCoaLikInf-class}
\alias{ecoCoaLikInf-class}
\title{Class to contain the likelihood inference}
\description{
Class to contain the likelihood inference
}
\section{Slots}{

\describe{
\item{\code{maxLik}}{matrix. This object contains the log likelihood of the coalescent simulations made under the parameters tested. Each column corresponds to the log likelihoods obtained for the simulated coalescents under \emph{one} parameter value. The number of columns should be equal to the number of parameter values tested.}

\item{\code{params}}{numeric. The parameter values under which the coalescent simulations were carried away.}
}}

