% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{ecoCoaLikInf}
\alias{ecoCoaLikInf}
\title{Creates an ecoCoaLikInf object}
\usage{
ecoCoaLikInf(ecoGenetSetL)
}
\arguments{
\item{ecoGenetSetL}{ecoGenetSetLik object. This object should contain the simulated coalescents for various parameter values together with their corresponding values.}
}
\value{
ecoCoalLikInf object. This object is used for the parameter inference.
}
\description{
This function uses the information contained in an ecoGenetSetLik object to create a matrix where the . This object is then used to infer the probability distribution of the parameter of interest.
}
