% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{ecoGenetSet-class}
\alias{ecoGenetSet-class}
\title{ecoGenetSet class}
\description{
This class reunites the genetic information of the samples contained in a genetSet object together with the spatial, demographic and social information contained in envDynSet.
@slot .Data envDynSet object containing all socioecological and demographic information.
@slot genetSample genetSet object containing all genetic and spatial information for the samples collected.
@export
}
