% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{ecoGenetSet}
\alias{ecoGenetSet}
\title{Creates an ecoGenetSet object}
\usage{
ecoGenetSet(envDyn = NULL, genetData = NULL)
}
\arguments{
\item{envDyn}{envDynSet object. The object containing the spatial, social and demographic information of the region under study.}

\item{genetData}{genetSet object. The object containing the genetic, temporal and location information of the collected samples.}
}
\value{
ecoGenetSet object.
}
\description{
This function creates an ecoGenetSet object from an envDynSet and a genetSet objects.
This is the preferred method for creating an ecoGenetSet object as opposed to \code{new("ecoGenetSet")} as this way some important calculations would not be performed.
}
