% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{ecogenetSetLik-class}
\alias{ecogenetSetLik-class}
\title{ecogenetSetLik class}
\description{
This class contains the different genealSimProb object generated with different values of a parameter that will be estimated using likelihood sampling.
}
\section{Slots}{

\describe{
\item{\code{genealSimProbList}}{list of genealSimProb objects that contain the socio-ecologic, demogenetic information, and the simulated genealogies for each object.}

\item{\code{likelihoodParams}}{vector containing the different values for the parameters on which the likelihood is performed.}
}}

