% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{ecogenetSetLik}
\alias{ecogenetSetLik}
\title{Creates a ecogenetSetLik object}
\usage{
ecogenetSetLik(genSimProb = NULL, likPar = NULL)
}
\arguments{
\item{genSimProb}{list of genealSimProb objects created with different values for the parameter of interest.}

\item{likPar}{vector of the parameter values corresponding to the different genealSimProb objects.}
}
\value{
ecogenetSetLik object.
}
\description{
This function creates an ecogenetSetLik object. This object is used for the inference of the probability distribution of a parameter from the likelihood sampling.
}
