% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{envDynSet-class}
\alias{envDynSet-class}
\title{envDynSet class}
\description{
This class calculates the forward and backwards socioecological transition matrices for the coalescence simulations starting from on the information present in the socioecoGeoDataModel object.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{socioecoGeoDataModel object containing the socioecological and geographical information of the studied area.}

\item{\code{RKlandscape}}{RasterStack object. R and K landscape calculated using the nicheModel objects corresponding to each environmental variable.}

\item{\code{geoDist}}{matrix. Distance matrix from each individual cell to each other in the raster.}

\item{\code{migrationMatrix}}{matrix. Calculated spatial migration matrix.}

\item{\code{transitionForward}}{matrix. Calculated forward transition matrix.}

\item{\code{transitionBackward}}{matrix. Calculated backwards transition matrix.}
}}

