% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{simulCoal}
\alias{simulCoal}
\alias{simulCoal,envDynSet,boolean}
\title{simulCoal method for envDynSet objects.}
\usage{
\S4method{simulCoal}{envDynSet,logical}(envDynSet, printCoal)
}
\description{
simulCoal method for envDynSet objects.
}
