% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{enveloppe}
\alias{enveloppe}
\title{Envelope reaction norm}
\usage{
enveloppe(X, p)
}
\arguments{
\item{X}{Numeric. Present value for the parameter.}

\item{p}{Numeric. Lower and higher limit for the reaction norm. \code{p} must have two values.}
}
\value{
Numeric. 1 for values within the lower and upper limits, and 0 for values outside.
}
\description{
An envelope reaction norm. Its value is 1 when \code{X} is within the values of the \code{p} parameter, and 0 when \code{X} is above or below the \code{p} values.
}
