% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{genealSimProb-class}
\alias{genealSimProb-class}
\title{genealSimProb class}
\description{
This class contains the ecoGenetSet information and a list of different coalescence simulations with its respective demographic and genetic probabilities.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{ecoGenetSet object containing the demographic and genetic information.}

\item{\code{genealogies}}{list of coalSim objects (simulated genealogies) with its corresponding probabilities.}
}}

