% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{genealSimProb}
\alias{genealSimProb}
\title{Creates a genealSimProb object}
\usage{
genealSimProb(ecoGenetSet = NULL, geneal = NULL)
}
\arguments{
\item{ecoGenetSet}{ecoGenetSet object. This contains all the environmental, demographic and genetic information required.}

\item{geneal}{list of coalSim simulations of the corresponding ecoGenetSet object. If \code{NULL}, an empty list will be created to store the coalescent simulations made from the ecoGenetSet object.}
}
\value{
genealSimProb
}
\description{
This function creates a genealSimProb object which contains the environmental, demographic, genetic information and the coalescent simulations together with its demographic and genetic probabilities.
}
