% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{genetData-class}
\alias{genetData-class}
\title{Class that contains the genetic data of the samples}
\description{
This class contains the spatial and temporal information (inherited from \link{spatialPoints} object) as well as the genetic data of the samples.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{spatialPoints object.}

\item{\code{type}}{Character.}

\item{\code{genetData}}{genotype object. information type.}

\item{\code{recombUnit}}{Recombination unit.}
}}

