% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{geneticDistance}
\alias{geneticDistance}
\title{Calculates the genetic transition probability between two markers}
\usage{
geneticDistance(x, y, mark, dist)
}
\arguments{
\item{x}{First marker to compare.}

\item{y}{Second marker to compare.}

\item{mark}{Character. Marker type. It can only be \code{microsatellite}.}

\item{dist}{Character. Genetic distance metric. Can be either \code{simple} or \code{geometric}.}
}
\value{
Double. Transition probability between two markers.
}
\description{
This function calculates the transition probability between two markers according to the type of marker and the distance metric selected.
}
