% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{geneticProb}
\alias{geneticProb}
\title{Genetic probability calculation for a genealogy}
\usage{
geneticProb(coalescent, ecoGenetData)
}
\arguments{
\item{coalescent}{coalSim object. This object should contain the simulated genealogy.}

\item{ecoGenetData}{ecoGenetSet object. This object should contain the genetic information of the samples.}
}
\value{
numeric. The calculated probability of the genealogy given the genetic characterization of the samples.
}
\description{
This function calculates the probability of a genealogy given the genetic data of a sample. This function requires a coalescent (simulated genealogy) and a ecoGenetSet object containing the marker information for the samples as well as the genetic transition matrix.
}
