% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{genetMatrix}
\alias{genetMatrix}
\alias{genetMatrix,genotype}
\alias{genetMatrix,genetSample}
\title{genetMatrix method for genotype objects.}
\usage{
\S4method{genetMatrix}{genotype}(object)

\S4method{genetMatrix}{genetSample}(object)
}
\description{
genetMatrix method for genotype objects.

genetMatrix method for genetSample objects.
}
