% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{genetMatrix}
\alias{genetMatrix}
\title{Method to generate a genetic allele matrix}
\usage{
genetMatrix(object)
}
\arguments{
\item{object}{Object that contains the genotypes of the samples.}
}
\value{
Matrix of the probabilities of the alleles in a sample of genotypes.
}
\description{
This method generates a genetic allele matrix from an object containing genotypes.
}
