% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{genetProb}
\alias{genetProb}
\title{Calculates the genetic probability of a genealogy}
\usage{
genetProb(sampleObj, coalSim)
}
\arguments{
\item{sampleObj}{object containing the genetic sample information.}

\item{coalSim}{genealogy from a coalescent simulation.}
}
\value{
numeric. Probability of a genealogy given the genetic information in the sample.
}
\description{
This function calculates the probability of a genealogy given the genetic data.
}
