% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{genetSample-class}
\alias{genetSample-class}
\title{Class that contains the genetic data of the samples}
\description{
This class contains the spatial and temporal information (inherited from \link{spatialPoints} object) as well as the genetic data of the samples.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{spatialPoints object.}

\item{\code{genetData}}{list of \code{genotype} objects. This list contains the information for the genetic markers of each of the sampled individuals.}

\item{\code{recombDist}}{matrix. This matrix contains the recombination probability between the studied markers. All marker names in the genotypes should also be in this recombination matrix.}
}}

