% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{genetSample}
\alias{genetSample}
\title{Creates a genetSample object.}
\usage{
genetSample(
  sampledPoints = NULL,
  samGenotypes = NULL,
  mType = NULL,
  recDist = NULL,
  mDist = TRUE
)
}
\arguments{
\item{sampledPoints}{samplePoints object containing the spatial information of the samples.}

\item{samGenotypes}{list of genotype objects corresponding to the sampled organisms.}

\item{mType}{character. Type of markers used. \code{microsatellite} is the name used for microsatellite markers.}

\item{recDist}{either a matrix or a data.frame containing the distance or the recombination probability between the markers.}

\item{mDist}{boolean. This indicates if the recombination matrix (or data.frame) contains recombination probabilities (\code{FALSE}) or distances in cM (\code{TRUE}). Defaults to \code{FALSE}.}
}
\value{
genetSample object.
}
\description{
This function creates a genetSample object.
}
