% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{genetSet-class}
\alias{genetSet-class}
\title{Class that includes genetSample, the mutation model and the gene distance matrix.}
\description{
This class includes inherits from genetSample the spatial and genetic data of the samples, and contains also the mutation model for the markers and the distance between the loci.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{genetSample object that contains the genetic and spatial information.}

\item{\code{mutationModel}}{Character. One of the implemented mutation models used to calculate the probability of coalescence.}
}}

