% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{genetSet}
\alias{genetSet}
\title{Creates a genetSet object}
\usage{
genetSet(genData = NULL, mutModel = NULL)
}
\arguments{
\item{genData}{genetSample object. This object should contain the genetic and spatial information of the sampled individuals.}

\item{mutModel}{mutationModel object. This specifies the mutation model selected for the calculation of the genetic transition probabilities.}
}
\value{
genetSet object.
}
\description{
This function creates a genetSet object. This uses a genetSample object and a mutation model object to calculate the genetic transition matrix.
}
