% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{genotype-class}
\alias{genotype-class}
\title{genotype class to hold the genetic information of individual samples}
\description{
This object holds the collection of loci that represents an individual.
}
\section{Slots}{

\describe{
\item{\code{loci}}{list of locus objects. This list contains the information of all the loci in the sample.}
}}

