% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{genotype}
\alias{genotype}
\title{Creates a genotype object.}
\usage{
genotype(samLoci = NULL)
}
\arguments{
\item{samLoci}{List of \code{locus} objects. All the objects on this list should be of the class \code{locus}.}
}
\value{
Object of the class \code{genotype}.
}
\description{
This function creates an object of genotype class.
}
