% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{geoEnvData-class}
\alias{geoEnvData-class}
\title{Class representing geographic and environmental data}
\description{
Class representing geographic and environmental data
}
\section{Slots}{

\describe{
\item{\code{layerConnectionTypes}}{character vector establishing the types of connection types of each layer. The amount of connection types should be the same as the amount of layers.}
}}

