% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{geoEnvData}
\alias{geoEnvData}
\title{Builder function for geoEnvData objects}
\usage{
geoEnvData(
  rasterStack = NULL,
  Array = array(c(c(5, 3, 2, 3, 2, 3, 2, 3, 5), rep(1:3, 3)), dim = c(3, 3, 2), dimnames
    = list(1:3, 1:3, c("temp", "pops"))),
  CRS = "+proj=longlat",
  xmn = 0,
  xmx = 3,
  ymn = 0,
  ymx = 3,
  layerConnectionTypes = c("geographic", "grouping")
)
}
\arguments{
\item{rasterStack}{a RasterStack object.}

\item{Array}{An array containing the values for the RasterStack layers.}
}
\description{
Builder function to create geoEnvData objects.
}
