% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{geoMigrationModel-class}
\alias{geoMigrationModel-class}
\title{Migration model based on the geographical distance or grouping variables.}
\description{
Proportion of individual produced in each attributed cell that migrates to each attributed cell of the raster.
models are described in successive components of modelConnectionType, varMig, shapeMig, pMig and pMixt slots
if modelConnectionType = dist and shapeMig=island the model is an island model in which each cell is an island and a proportion of propagules migrates to the pool with a proportion given by the parameter.
if modelConnectionType = group and shapeMig=island the model is an island model in which each group is an island and a proportion of propagules migrates to the pool with a proportion given by the parameter there is a possibility of dispersion models within groups according to the shapeDisp.
Models are described in successive components of modelConnectionType, varMig, shapeMig, pMig and pMixt slots.
}
\section{Slots}{

\describe{
\item{\code{modelConnectionType}}{Character. describes migration model: depending on distance or grouping.}

\item{\code{varMig}}{Variable according to which the individual migrates.}

\item{\code{shapeMig}}{Character. Type of probability distribution for the dispersal. The popSep model means island model where cells of the raster belong to pops according to their group variable and mix panmictically within groups according to the popSep model.}

\item{\code{pMixt}}{Numeric. A proportion of the individual migrates according to each model.}

\item{\code{pMig}}{List. Parameters of the corresponding probability distribution.}
}}

