% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{geoMigrationModel}
\alias{geoMigrationModel}
\title{Builder function for geoMigrationModel objects.}
\usage{
geoMigrationModel(
  modelConnectionType = c("geographic", "grouping"),
  varMig = c("temp", "pops"),
  shapeMig = c("gaussian", "island"),
  pMig = list(gaussian = 1/1.96, island = 0.2),
  pMixt = c(0.5, 0.5)
)
}
\arguments{
\item{modelConnectionType}{Character. describes migration model: depending on distance or grouping.}

\item{varMig}{Variable according to which the individual migrates.}

\item{shapeMig}{Character. Type of probability distribution for the dispersal. The popSep model means island model where cells of the raster belong to pops according to their group variable and mix panmictically within groups according to the popSep model.}

\item{pMig}{List. Parameters of the corresponding probability distribution.}

\item{pMixt}{Numeric. A proportion of the individual migrates according to each model.}
}
\value{
Object of the class geoMigrationModel.
}
\description{
Builder function for objects of the geoMigrationModel class.
}
