% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{getAlleles}
\alias{getAlleles}
\alias{getAlleles,genotype}
\title{getAlleles method for genotype objects.}
\usage{
\S4method{getAlleles}{genotype}(object)
}
\description{
getAlleles method for genotype objects.
}
