% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{getAlleles}
\alias{getAlleles}
\title{Generic method getAlleles}
\usage{
getAlleles(object)
}
\arguments{
\item{object}{\code{genotype} or \code{haplotype} class object. This method returns the allele values for each marker in the genotype or haplotype.}
}
\value{
Numeric list.
}
\description{
This method obtains all the alleles in a genotype or haplotype.
}
