% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{getGenealProb}
\alias{getGenealProb}
\alias{getGenealProb,genealSimProb}
\title{getGenealProb method for genealSimProb objects.}
\usage{
\S4method{getGenealProb}{genealSimProb}(object)
}
\description{
getGenealProb method for genealSimProb objects.
}
