% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{getGenealProb}
\alias{getGenealProb}
\title{Print the demographic and genetic probabilities of an object}
\usage{
getGenealProb(object)
}
\arguments{
\item{object}{the object containing the simulations.}
}
\value{
data frame with the demographic and genetic probabilities.
}
\description{
This function returns a data frame with the demographic and genetic probabilities for the genealigy simulations contained in an object.
}
