% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{getHaplotypes}
\alias{getHaplotypes}
\alias{getHaplotypes,genotype}
\title{getHaplotypes method for genotype objects.}
\usage{
\S4method{getHaplotypes}{genotype}(object, num = NULL)
}
\description{
getHaplotypes method for genotype objects.
}
