% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{getHaplotypes}
\alias{getHaplotypes}
\title{Generic method getHaplotypes}
\usage{
getHaplotypes(object, num = NULL)
}
\arguments{
\item{object}{\code{genotype} class object. From this object the haplotypes are sampled.}
}
\value{
haplotype objects.
}
\description{
This method obtains all or the specified number of possible haplotypes that a genotype could produce.
}
