% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{getPloidy}
\alias{getPloidy}
\alias{getPloidy,locus}
\alias{getPloidy,genotype}
\title{getPloidy method for locus objects.}
\usage{
\S4method{getPloidy}{locus}(object)

\S4method{getPloidy}{genotype}(object)
}
\description{
This method returns the ploidy level for the objects of class locus.

This method returns the ploidy level for each of the markers in a genotype object.
}
