% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{getPloidy}
\alias{getPloidy}
\title{Generic method for to obtain the ploidy of a genotype.}
\usage{
getPloidy(object)
}
\arguments{
\item{object}{object with genetic markers.}
}
\value{
Int. Ploidy level of the marker(s).
}
\description{
This method returns the ploidy for a marker or a genotype.
}
