\name{GraphPOP-package}
\alias{GraphPOP-package}
\alias{GraphPOP}
\docType{package}
\title{
\packageTitle{GraphPOP}
}
\description{
\packageDescription{GraphPOP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GraphPOP}
\packageIndices{GraphPOP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GraphPOP}

Maintainer: \packageMaintainer{GraphPOP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
