% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{haplotype-class}
\alias{haplotype-class}
\title{Class that contains a single haplotype}
\description{
An object of this class contains represents a single haplotype. For this reason it cannot have more than one allele at each locus.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{Genotype object. This is a class that inherits from the \code{genotype} with the constrain that it can only have one allele per locus.}
}}

