% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{haplotype}
\alias{haplotype}
\title{Creates a haplotype object}
\usage{
haplotype(samLoci = NULL)
}
\arguments{
\item{samLoci}{List of \code{locus} objects. All the objects on this list should be of the class \code{locus}. For \code{haplotype} objects there must be only one allele per locus.}
}
\value{
Object of the haplotype class.
}
\description{
This function creates a haplotype object.
}
