% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{hypCi}
\alias{hypCi}
\alias{hypCi,ecoCoaLikInf}
\title{hypCi method for ecoCoaLikInf objects}
\usage{
\S4method{hypCi}{ecoCoaLikInf}(object, level = 0.95, ci = "MESLE")
}
\arguments{
\item{object}{ecoCoaLikInf object. This contains the parameter values, the simulations done with different parameter values and the log-likelihood of each simulation.}

\item{level}{A numeric vector of confidence levels.}

\item{ci}{A character string indicating the quantity for which a confidence interval is to be constructed. Either "MESLE" or "parameter". See Details.}
}
\description{
hypCi method for ecoCoaLikInf objects
}
\seealso{
\link[sbi:ci]{sbi::ci}
}
