% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{hypCi}
\alias{hypCi}
\title{One-dimensional confidence interval for the parameter estimation}
\usage{
hypCi(object, level = 0.95, ci = "MESLE")
}
\arguments{
\item{object}{Object containing the simulations.}

\item{level}{A numeric vector of confidence levels.}

\item{ci}{A character string indicating the quantity for which a confidence interval is to be constructed. Either "MESLE" or "parameter". See Details.}
}
\value{
confidence interval.
}
\description{
This method builds a one-dimensional confidence interval for the MLE estimation of the parameter of interest.
}
