% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{hypTest}
\alias{hypTest}
\alias{hypTest,ecoCoaLikInf}
\title{hypTest method for ecoCoaLikInf object.}
\usage{
\S4method{hypTest}{ecoCoaLikInf}(
  object,
  null.value,
  test = c("parameter", "MESLE", "moments"),
  case = NULL,
  type = NULL,
  weights = NULL,
  K1_est_method = "batch",
  batch_size = NULL,
  max_lag = NULL,
  plot_acf = FALSE,
  MCcorrection = "none",
  ...
)
}
\arguments{
\item{object}{ecoCoaLikInf object. This contains the parameter values, the simulations done with different parameter values and the log-likelihood of each simulation.}

\item{null.value}{The null value(s) for the hypothesis test. The expected format depends on which teset will be carried out. See the Details section for more information.}

\item{test}{A character string indicating which is to be tested about. One of "moments", "MESLE", or "parameter". See Details.}

\item{case}{When \code{test} is "parameter", \code{case} needs to be either "iid" or "stationary". \code{case} = "iid" means that the observations are iid, and \code{case} = "stationary" means that the observations form a stationary sequence. The \code{case} argument affects how the variance of the slope of the mean function (=K_1 in Park (2023)) is estimated. The default value is "stationary".}

\item{type}{When \code{test} is "moments", the \code{type} argument needs to be specified. \code{type} = "point" means that the test about the mean and the variance of simulation log likelihoods at a given parameter point is considered. \code{type} = "regression" means that the test about the mean function and the variance of simulation log likelihoods at various parameter values is considered. See Details.}

\item{weights}{An optional argument. The un-normalized weights of the simulation log likelihoods for regression. A numeric vector of length equal to the \code{params} attribute of the \code{simll} object. See Details below.}

\item{K1_est_method}{Either "batch" or "autocov". Used when \code{test} is "parameter" and \code{case} is "stationary". The default is "batch". See Details for more information.}

\item{batch_size}{Numeric. The size of the batch when \code{K1_est_method} is "batch". If not supplied, the default value is \code{round(n^0.4)} where \code{n} is the number of observations in the data.}

\item{max_lag}{When \code{test} is "parameter" and \code{case} is "stationary", the value of \code{max_lag} gives the truncation point for lagged autocovariance when estimating K1 as a sum of lagged autocovariances of estimates slopes. If not supplied, default is the maximum lag for which at least one of the entries of the matrix of lagged autocorrelation has absolute value greater than 4/sqrt(nobs), where the lagged autocorrelation is found up to lag \code{10*log10(nobs/d)}. Here \code{nobs} is the number of observations and \code{d} is the dimension of the parameter space.}

\item{plot_acf}{Logical. Should the autocorrelation plot be generated when estimating K1 for the case where \code{test} is "parameter" and \code{case} is "stationary"?}

\item{MCcorrection}{For tests on the simulation based parameter surrogate (\code{test}="parameter"), \code{MCcorrection} determines if and how the sampling distribution of the test statistic will be corrected by a Monte Carlo method to account for the variability in the estimate of K1. Possible values are "none" (default) and "Wishart". See the Details section and Park (2023) for more details.}

\item{...}{Other optional arguments, not currently used.}
}
\description{
hypTest method for ecoCoaLikInf object.
}
\seealso{
\link[sbi:ht]{sbi::ht}
}
