% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{hypTest}
\alias{hypTest}
\title{hypTest generic method}
\usage{
hypTest(
  object,
  null.value,
  test = c("parameter", "MESLE", "moments"),
  case = NULL,
  type = NULL,
  weights = NULL,
  K1_est_method = "batch",
  batch_size = NULL,
  max_lag = NULL,
  plot_acf = FALSE,
  MCcorrection = "none",
  ...
)
}
\arguments{
\item{object}{The object containing the simulations to maximize the likelihood.}

\item{...}{Additional parameters to be passed on to \link{ht}.}
}
\value{
sbim::ht object.
}
\description{
Used for hypothesis testing for models defined implicitly by a random simulator as described by \href{https://arxiv.org/abs/2311.09446}{Park 2025}.
}
