% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{locus-class}
\alias{locus-class}
\title{locus class to contain individual locus information}
\description{
The \code{locus} class is used to store the information of a particular locus. It contains the locus name and the allele values for either microsatellite markers or presence/absence.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character. The name of the locus.}

\item{\code{markerType}}{Character. The type of marker used. This is important for the selection of an appropriate mutation model.}

\item{\code{alleles}}{Numeric. Vector that contains the allele values.}
}}

