% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{locus}
\alias{locus}
\title{Create a locus object}
\usage{
locus(locusName, mType, mValues)
}
\arguments{
\item{locusName}{Character. The name of the marker.}

\item{mType}{Character. The type of marker used. This class accepts \code{microsatellite} and \code{SNP} data.}

\item{mValues}{Numeric. The value(s) of the marker alleles.}
}
\value{
a locus object.
}
\description{
This function creates a locus object.
}
