% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{markerCompare}
\alias{markerCompare}
\title{Compare genetic markers}
\usage{
markerCompare(x, marker, distMet)
}
\arguments{
\item{x}{the markers to be compared.}

\item{marker}{the type of marker used.}

\item{distMet}{the distance metric used.}
}
\value{
a probability transition matrix.
}
\description{
This function compares genetic markers after the genetic matrix is done.
}
