% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{markerDist}
\alias{markerDist}
\alias{markerDist,locus}
\alias{markerDist,genotype}
\title{markerDist method for locus objects.}
\usage{
\S4method{markerDist}{locus,locus}(objectA, objectB)

\S4method{markerDist}{genotype,genotype}(objectA, objectB)
}
\description{
This method returns the step distance between two representatives of the same locus.

This method returns the step distance for all markers in a pair of genotype objects.
}
