% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{markerDist}
\alias{markerDist}
\title{Method to retrieve the distance between two loci.}
\usage{
markerDist(objectA, objectB)
}
\arguments{
\item{objectA}{First locus to compare.}

\item{objectB}{Second locus to compare.}
}
\value{
Int. Steps distance between loci.
}
\description{
This method returns the steps distance between pairs of loci. The loci must have the same name.
}
