% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{markerNames}
\alias{markerNames}
\alias{markerNames,genotype}
\title{markerNames method for genotype objects.}
\usage{
\S4method{markerNames}{genotype}(object)
}
\description{
markerNames method for genotype objects.
}
