% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{markerNames}
\alias{markerNames}
\title{Generic method markerNames}
\usage{
markerNames(object)
}
\arguments{
\item{object}{\code{genotype} or \code{haplotype} class object. This method returns the names for each marker in the genotype or haplotype.}
}
\value{
Character containing the names of the markers.
}
\description{
This method obtains all the marker names in a genotype or haplotype.
}
