% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{markerTypes}
\alias{markerTypes}
\alias{markerTypes,genotype}
\title{markerTypes method for genotype objects.}
\usage{
\S4method{markerTypes}{genotype}(object)
}
\description{
markerTypes method for genotype objects.
}
