% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{markerTypes}
\alias{markerTypes}
\title{Generic method markerTypes}
\usage{
markerTypes(object)
}
\arguments{
\item{object}{\code{genotype} or \code{haplotype} class object. This method returns the type of each marker in the genotype or haplotype.}
}
\value{
Character containing the type of the markers.
}
\description{
This method obtains all the marker types in a genotype or haplotype.
}
