% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{model}
\alias{model}
\alias{model,nicheModel}
\title{model method for nicheModel objects.}
\usage{
\S4method{model}{nicheModel}(object)
}
\description{
model method for nicheModel objects.
}
