% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{model.extract}
\alias{model.extract}
\alias{model.extract,nicheModel}
\alias{model.extract,socioecoMigrationModel}
\title{model.extract method for nicheModel objects.}
\usage{
\S4method{model.extract}{nicheModel}(frame, component = "reactNorms")

\S4method{model.extract}{socioecoMigrationModel}(frame, component = "shapeMig")
}
\description{
model.extract method for nicheModel objects.

model.extract method for socioecoMigrationModel objects.
}
