% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{model}
\alias{model}
\title{Obtain the niche model of an object.}
\usage{
model(object)
}
\arguments{
\item{object}{Object to extract the model.}
}
\value{
nicheModel object.
}
\description{
This function obtains the niche model of an object.
}
