% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{mutationModel-class}
\alias{mutationModel-class}
\title{mutationModel class}
\description{
This class holds the mutation model used for calculating the genetic transition matrix. This only provides some barriers to coerce the possible values of the mutation models.
}
