% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{mutationModel}
\alias{mutationModel}
\title{Creates a mutationModel object}
\usage{
mutationModel(x)
}
\arguments{
\item{x}{character string indicating the mutation models. Currently only \code{simple} and \code{geometric} are supported.}
}
\value{
mutationModel object.
}
\description{
This function creates a mutationModel object.
}
