% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{NAcells}
\alias{NAcells}
\alias{NAcells,geoEnvData}
\alias{Acells}
\alias{Acells,geoEnvData}
\title{NA cells.}
\usage{
\S4method{NAcells}{geoEnvData}(object)

\S4method{Acells}{geoEnvData}(object)
}
\description{
NA cells.

A cells.
}
