% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{NAcells}
\alias{NAcells}
\title{Identify NA values}
\usage{
NAcells(object)
}
\arguments{
\item{object}{a raster object to evaluate the values.}
}
\value{
vector of cells which value is NA.
}
\description{
\code{NAcells} returns the cells with NA values in a raster object.
}
