% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{nCellA}
\alias{nCellA}
\title{Counts the number of non-NA cells}
\usage{
nCellA(object)
}
\arguments{
\item{object}{raster object.}
}
\value{
integer counting the amount of non-NA cells.
}
\description{
This function returns the number of cells in a raster containing non-NA values.
}
