% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{nicheModel-class}
\alias{nicheModel-class}
\title{nicheModel class}
\description{
Defines the reaction norm for each variable, without interaction in this version.
}
\section{Slots}{

\describe{
\item{\code{varNiche}}{Character. The variable names to which the niche model applies.}

\item{\code{reactNorms}}{Character. The reaction norms for the variables as a vector. The names of the entries correspond to the variables the reaction norm uses, and the value corresponds to the type of reaction norm.}

\item{\code{pNiche}}{List. The list of numerical parameters vectors for each reaction norm, there should be a parameter vector for each reaction norm. The names of the vector components correspond to the values of the reaction norm slot.}
}}

