% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{nicheModel}
\alias{nicheModel}
\title{nicheModel builder function}
\usage{
nicheModel(
  varNiche = c("temp", "temp"),
  reactNorms = c("envelin", "scaling"),
  pNiche = list(c(3, 4), 100)
)
}
\arguments{
\item{varNiche}{Character. The variable names to which the niche model applies.}

\item{reactNorms}{Character. The reaction norms for the variables as a vector. The names of the entries correspond to the variables the reaction norm uses, and the value corresponds to the type of reaction norm.}

\item{pNiche}{List. The list of numerical parameters vectors for each reaction norm, there should be a parameter vector for each reaction norm. The names of the vector components correspond to the values of the reaction norm slot.}
}
\value{
an object of nicheModel class.
}
\description{
Builder function for nicheModel objects.
}
