% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{npMig}
\alias{npMig}
\title{Function to return the amount of parameters for a migration kernel.}
\usage{
npMig(x)
}
\arguments{
\item{x}{Character. The name of the migration probability distribution.}
}
\value{
Int. Returns the amount of parameters for the input probability distribution.
}
\description{
Function used to return the amount of parameters of a migration probability distribution.
}
