% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{npNiche}
\alias{npNiche}
\title{Function for the amount of parameters of a niche model.}
\usage{
npNiche(x)
}
\arguments{
\item{x}{Character. The type of the niche model.}
}
\value{
Int. Number of parameters for the niche model.
}
\description{
Returns the amount of parameters required for determined niche model.
}
