% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{plot.coalescent}
\alias{plot.coalescent}
\alias{plot}
\title{Plot method for coalescent objects}
\usage{
\method{plot}{coalescent}(object)
}
\arguments{
\item{object}{coalescent. The coalescent object.}
}
\value{
plot of the tree representing the coalescent.
}
\description{
Plot method for coalescent objects.
}
