% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{print.coalescent}
\alias{print.coalescent}
\alias{print}
\title{Print method for coalescent objects}
\usage{
\method{print}{coalescent}(object)
}
\arguments{
\item{object}{coalescent. The coalescent object.}
}
\value{
plot of the tree representing the coalescent.
}
\description{
Print method for coalescent objects.
}
