% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{sampledCells-class}
\alias{sampledCells-class}
\title{Class that contains the information of the sampled cells.}
\description{
This class holds the geographical (coordinates), temporal (times of collection), and socioecological characteristics of the samples taken.
}
\section{Slots}{

\describe{
\item{\code{geoCoordinates}}{SpatialPoints object. Coordinates for the samples.}

\item{\code{sampleCell}}{Int. Cells where the samples are located}

\item{\code{sampleTime}}{Int. Times when the samples where collected. These times are used to integrate the past samples into the coalescence simulations.}

\item{\code{socioecoCoordinates}}{Character list. Socioecological coordinates for the samples.}
}}

