% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{samplePoints}
\alias{samplePoints}
\title{Create objects of class samplePoints}
\usage{
samplePoints(sCoordinates, sTimes, proj4 = NULL)
}
\arguments{
\item{sCoordinates}{SpatialPoints, matrix or data.frame object. Coordinates where the samples were collected. In case of not being a \code{SpatialPoints} object, the data.frame or matrix should have only two columns, for Latitude and Longitude respectively.}

\item{sTimes}{Sampling times.}

\item{proj4}{Projection string of class \link[sp]{CRS}. Default value \code{NULL}. If NULL, \code{+proj=longlat +datum=WGS84 +no_defs} is used.}
}
\value{
samplePoints object.
}
\description{
Creates an object of the class samplePoints
}
