% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{scaling}
\alias{scaling}
\title{Scaling function}
\usage{
scaling(X, p)
}
\arguments{
\item{X}{Numeric. Reaction norm value.}

\item{p}{Numeric. Scaling parameter.}
}
\value{
Scaled value.
}
\description{
Function used to scale the reaction norms.
}
