% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{setName}
\alias{setName}
\alias{setName,locus}
\title{setName method for locus objects.}
\usage{
\S4method{setName}{locus}(object, newName = NULL)
}
\description{
setName method for locus objects.
}
