% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{setName}
\alias{setName}
\title{Generic method setName}
\usage{
setName(object, newName = NULL)
}
\arguments{
\item{object}{\code{locus} class object. This method changes the name of a marker in a locus.}
}
\value{
locus object
}
\description{
Sets the genetic marker name of a locus.
}
