% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{setType}
\alias{setType}
\alias{setType,locus}
\title{setType method for locus objects.}
\usage{
\S4method{setType}{locus}(object, newType = NULL)
}
\description{
setType method for locus objects.
}
