% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{setType}
\alias{setType}
\title{Generic method setType}
\usage{
setType(object, newType = NULL)
}
\arguments{
\item{object}{\code{locus} class object. This method changes the type of a marker in a locus.}
}
\value{
locus object.
}
\description{
Sets the genetic marker type of a locus.
}
