% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{simulCoal}
\alias{simulCoal}
\alias{simulCoal,genealSimProb,boolean}
\alias{#'}
\alias{@description}
\alias{Simulates}
\alias{a}
\alias{coalescence}
\alias{from}
\alias{socioecological}
\alias{and}
\alias{geographical}
\alias{data.}
\title{simulCoal method for genealSimProb objects.}
\usage{
\S4method{simulCoal}{genealSimProb,logical}(ecoGenetSet, printCoal)
}
\arguments{
\item{ecoGenetSet}{genealSimProb object. This object must contain all the information for the coalescence simulation.}

\item{printCoal}{Boolean. If \code{TRUE} it prints the time elapsed during the simulations.}
}
\value{
List. The first element is the coalescent, and the second is the summed forward probability.
}
\description{
simulCoal method for genealSimProb objects.
}
