% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{simulCoal}
\alias{simulCoal}
\title{Coalescence simulation from socioecological and geographical data.}
\usage{
simulCoal(envDynSet, printCoal)
}
\arguments{
\item{envDynSet}{envDynSet object. This object must contain all the information for the coalescence simulation.}

\item{printCoal}{Boolean. If \code{TRUE} it prints the time elapsed during the simulations.}
}
\value{
List. The first element is the coalescent, and the second is the summed forward probability.
}
\description{
Simulates a coalescence from socioecological and geographical data.
}
