% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{simulMultiCoal}
\alias{simulMultiCoal}
\alias{simulMultiCoal,ecoGenetSet}
\title{simulMultiCoal method for ecoGenetData objects.}
\usage{
\S4method{simulMultiCoal}{ecoGenetSet,logical,numeric}(ecoGenetSet, printCoal, iteration)
}
\description{
simulMultiCoal method for ecoGenetData objects.
}
