% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{simulMultiCoal}
\alias{simulMultiCoal}
\title{Method to simulate multiple coalSim.}
\usage{
simulMultiCoal(ecoGenetSet, printCoal, iteration, multiCore, cores)
}
\arguments{
\item{ecoGenetSet}{ecoGenetSet object that contains the spatial, demographic and genetic information.}

\item{printCoal}{Boolean. \code{TRUE} if the time of each simulation should be displayed.}

\item{iteration}{Int. The amount of simulations to perform.}
}
\value{
list of coalSim objects.
}
\description{
This method is used to simulate multiple coalescents for parameter inference.
}
