% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{socioecoGeoData-class}
\alias{socioecoGeoData-class}
\title{Class to contain the socio-ecological and the geographical data.}
\description{
socioecoGeoData are raster layers containing information about environmental variables and or connectivity variables that will allow to build niche models and friction models
the layers that contain geographic information related to geographic coordinates are called connectionType = geographic
the layers that contain other type of information characterizing ecological populations by groups (ethnic groups, market exchanges, plant varieties for pests and diseases) have a connectionType = grouping
connected class variables are coded as follows; when cell value is:
\itemize{
\item 0 : the cell is not connected to any other cell
\item n!=0 : the cell is connected to the cells having the same value
}
}
\section{Slots}{

\describe{
\item{\code{geoEnvData}}{geoEnvData object. This object contains the geographical data.}

\item{\code{socioecoData}}{socioecoGroupsData object. This object contains the data for the socio-ecological grouping data.}
}}

