% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{socioecoGeoData}
\alias{socioecoGeoData}
\title{Creates a socioecoGeoData object}
\usage{
socioecoGeoData(
  x = NULL,
  socioecoList = NULL,
  stackConnectionType = NULL,
  envLayerNames = NULL
)
}
\arguments{
\item{x}{Either a geoEnvData or a rasterStack object containing the geographical information.}

\item{socioecoList}{socioecoGroupsData object containing the groups and connection types.}

\item{stackConnectionType}{character. The name of the stack connection type. Default value \code{geographic}.}
}
\value{
socioecoGeoData object.
}
\description{
Function to construct a socioecoGeoData object.
}
