% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{socioecoGeoDataHistory}
\alias{socioecoGeoDataHistory}
\title{Creates an object of the class socioecoGeoDataHistory.}
\usage{
socioecoGeoDataHistory(
  SocioecoGeoData = socioecoGeoData(),
  PastSocioecoGeoData = list(socioecoGeoData(), socioecoGeoData(), socioecoGeoData()),
  ParsingTimes = c(0, -200, -500, -2000),
  TimeUnit = "days",
  ZeroTime = as.POSIXlt("2005-4-19 7:01:00")
)
}
\arguments{
\item{PastSocioecoGeoData}{List of socioecoGeoData representing the past socio-ecological and geographical data.}

\item{ParsingTimes}{Numeric. List of the parsing times.}

\item{TimeUnit}{Character. Specifies the time unit. Defaults to "days".}

\item{ZeroTime}{POSIXlt. The present time.}

\item{socioecoGeoData}{socioecoGeoData object that contains the present data.}
}
\value{
An object of the socioecoGeoDataHistory class.
}
\description{
Class to represent environmental dynamic data history. It inherits from socioecoData as the present socioecogeodata. Includes a past socioecogeodata list with parsing times.
The last past socioecogeodata in the list goes from the last parsing time to minus infinite.
}
