% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{socioecoGeoDataHistoryAndSample-class}
\alias{socioecoGeoDataHistoryAndSample-class}
\title{Class that reunites the present and past socioecoData.}
\description{
Class to represent environmental dynamic data history. It inherits from socioecoData as the present socioecogeodata. Includes a past socioecogeodata list with parsing times.
The last past socioecogeodata in the list goes from the last parsing time to minus infinite.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{socioecoGeoData object that contains the present data.}

\item{\code{pastSocioecoGeoData}}{List of socioecoGeoData representing the past socio-ecological and geographical data.}

\item{\code{parsingTimes}}{Numeric. List of the parsing times.}

\item{\code{timeUnit}}{Character. Specifies the time unit.}

\item{\code{zeroTime}}{POSIXlt. The present time.}
}}

