% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{socioecoGeoDataModel-class}
\alias{socioecoGeoDataModel-class}
\title{Class to contain the socioecoGeoDataModel built on the historical information}
\description{
This class builds the migration models using the historical and present socioecological and geographical information.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{socioecoGeoDataHistory object containing all the historical and present socioecological and geographical information.}

\item{\code{SocioecoGeoData}}{socioecoGeoData object.}

\item{\code{Kmodel}}{nicheModel object. Carrying capacity niche model.}

\item{\code{Rmodel}}{nicheModel object. Reproductive potential niche model.}

\item{\code{geoMigModel}}{geoMigrationModel object. Contains the geographic migration model information.}

\item{\code{socioecoMigModel}}{socioecoMigrationModel object. Contains the socioecological migration model information.}

\item{\code{sampledPoints}}{samplePoints object. Contains the coordinates of the samples, the sampling times and the cells where these samples are located.}
}}

