% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{socioecoGeoDataModel}
\alias{socioecoGeoDataModel}
\title{Create a socioecoGeoDataModel object.}
\usage{
socioecoGeoDataModel(
  socioecoGeoDataHistory = NULL,
  SocioecoGeoData = socioecoGeoData(),
  PastSocioecoGeoData = list(socioecoGeoData(), socioecoGeoData(), socioecoGeoData()),
  ParsingTimes = c(0, -200, -500, -2000),
  TimeUnit = "days",
  ZeroTime = as.POSIXlt("2005-4-19 7:01:00"),
  nicheK = NULL,
  nicheR = NULL,
  migModel = NULL,
  varNicheK = "temp",
  reactNormsK = c(temp = "scaling"),
  pNicheK = list(scalingK = 100),
  varNicheR = c("temp", "temp"),
  reactNormsR = c(temp = "envelin", temp = "scaling"),
  pNicheR = list(envelin = c(1, 4), scalingR = 10),
  modelConnectionType = c("geographic", "grouping"),
  varMig = c("temp", "pops"),
  shapeMig = c("gaussian", "popSep"),
  pMig = list(110574/1.96, numeric(0)),
  pMixt = c(0.5, 0.5)
)
}
\arguments{
\item{socioecoGeoDataHistory}{socioecoGeoDataHistory object. This object contains the socio-ecological, geographical and historical data for the studied population. Default value \code{NULL}.}

\item{SocioecoGeoData}{socioecoGeoData object. If the socioecoGeoDataHistory object is missing, this parameter contains the present socio-ecological and geographical data.}

\item{PastSocioecoGeoData}{List of socioecoGeoData objects. If the socioecoGeoDataHistory object is missing, this parameter contains the past socio-ecological and geographical data.}

\item{ParsingTimes}{Numeric. Parsing times for the present and past data.}

\item{TimeUnit}{Character. The time unit for the Parsing Times.}

\item{ZeroTime}{POSIXlt object. The zero or present time.}

\item{nicheK}{nicheModel object. Niche model for the carrying capacity.}

\item{nicheR}{nicheModel object. Niche model for the reproductive capacity.}

\item{migModel}{migrationModel object. The migration model calculated for the studied landscape.}

\item{varNicheK}{Character. The variable name of which the carrying capacity K depends.}

\item{reactNormsK}{Character. The type of reaction norm for the listed variable for the carrying capacity.}

\item{pNicheK}{Numeric. The parameter(s) for the reaction norm of the carrying capacity (K).}

\item{varNicheR}{Character. The variable name of which the reproductive potential R depends.}

\item{reactNormsR}{Character. The type of reaction norm for the listed variable for the reproductive potential.}

\item{pNicheR}{Numeric. The parameter(s) for the reaction norm of the reproductive potential (R).}

\item{modelConnectionType}{Character. The types of connection types in the model.}

\item{varMig}{Character. The variables involved in the migration model calculation.}

\item{shapeMig}{Character. The shape of the migration probability distribution.}

\item{pMig}{Numeric. Migration probability.}

\item{pMixt}{Numeric. Mixture probability.}
}
\value{
socioecoGeoDataModel object.
}
\description{
Create a socioecoGeoDataModel object.
}
