% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{socioecoGroupData-class}
\alias{socioecoGroupData-class}
\title{Class representing the socio-economic grouping data}
\description{
Class representing the socio-economic grouping data
}
\section{Slots}{

\describe{
\item{\code{layerConnectionTypes}}{}

\item{\code{categories}}{character. Vector establishing the types of connection types of each layer. The amount of connection types should be the same as the amount of layers.}
}}

