% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{socioecoGroupData}
\alias{socioecoGroupData}
\title{Creates a socioecoGroupData object}
\usage{
socioecoGroupData(
  categories = c("group1", "group2"),
  Values = c(1:4),
  Nlayers = 2,
  layerNames = c("cuidado", "alimentación"),
  Array = NULL,
  rasterStack = NULL
)
}
\arguments{
\item{categories}{character. The categories of the socio-ecological groups.}

\item{Values}{numeric. Values for the rasterStack.}

\item{Array}{array. Array with the values of the socio-ecological groups per cell.}

\item{rasterStack}{rasterStack object. This object contains the ordered values within a rasterStack object.}
}
\value{
socioecoGroupData object.
}
\description{
This function creates a socioecoGroupData object.
}
