% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{socioecoMigrationModel-class}
\alias{socioecoMigrationModel-class}
\title{Class to describe a migration model given by the socioecological variables.}
\description{
Proportion of individual produced in each attributed cell that migrates to each attributed cell of the raster.
}
\section{Slots}{

\describe{
\item{\code{weight}}{Numeric. represents the coefficient given to socioecoMigration Matrix relative to geoMigration matrix (has a fixed weight of 1).}

\item{\code{varMig}}{Character. represents the names of the socioecoGroupsData used to calculate migration rates.}

\item{\code{shapeMig}}{represents the model used to transform varMig into migration rates.}

\item{\code{pMig}}{represent the parameters to transform varMig into distance.}
}}

