% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{socioecoMigrationModel}
\alias{socioecoMigrationModel}
\title{Builder function for a socioecoMigrationModel object.}
\usage{
socioecoMigrationModel(
  weight = 1,
  varMig = c("Español", "Chibcha", "Corrabastos", "Semillas_Comerciales",
    "Guardianes_de_semilla"),
  shapeMig = c("euclideanInverse"),
  pMig = list(Español = 1, Chibcha = 1, Corrabastos = 1, Semillas_Comerciales = 1,
    Guardianes_de_semilla = 1)
)
}
\arguments{
\item{weight}{Numeric. represents the coefficient given to socioecoMigration Matrix relative to geoMigration matrix (has a fixed weight of 1).}

\item{varMig}{Character. represents the names of the socioecoGroupsData used to calculate migration rates.}

\item{shapeMig}{represents the model used to transform varMig into migration rates.}

\item{pMig}{represent the parameters to transform varMig into distance.}
}
\value{
An object of the socioecoMigrationModel class.
}
\description{
Proportion of individual produced in each attributed cell that migrates to each attributed cell of the raster.
}
