% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{[,ecoGenetSet,character,missing,ANY-method}
\alias{[,ecoGenetSet,character,missing,ANY-method}
\title{Extracts data from ecoGenetSet objects}
\usage{
\S4method{[}{ecoGenetSet,character,missing,ANY}(x, i, j, drop)
}
\description{
Extracts data from ecoGenetSet objects
}
