% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{[,envDynSet,character,missing,ANY-method}
\alias{[,envDynSet,character,missing,ANY-method}
\title{Extracts data from envDynSet objects}
\usage{
\S4method{[}{envDynSet,character,missing,ANY}(x, i, j, drop)
}
\description{
Extracts data from envDynSet objects
}
