% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{TransitionBackward-class}
\alias{TransitionBackward-class}
\title{Backward transition probabilities matrix}
\description{
Square matrix containing the backward transition probabilities of a defined landscape.
}
