% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{TransitionBackward}
\alias{TransitionBackward}
\title{Create a TransitionBackward object}
\usage{
TransitionBackward(matrix)
}
\arguments{
\item{matrix}{Matrix containing the backwards transition probability.}
}
\value{
TransitionBackward object.
}
\description{
This function creates a TransitionBackward object.
}
