% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{class}
\name{TransitionForward-class}
\alias{TransitionForward-class}
\title{Forward transition probabilities matrix}
\description{
Square matrix containing the forward transition probabilities of a defined landscape.
}
