% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{validitygeoMigrationModel}
\alias{validitygeoMigrationModel}
\title{Validity function for geoMigrationModel objects.}
\usage{
validitygeoMigrationModel(object)
}
\arguments{
\item{object}{geoMigrationModel object.}
}
\value{
boolean.
}
\description{
Validity function for geoMigrationModel objects.
}
