% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{validityLocus}
\alias{validityLocus}
\title{Validity function for locus class}
\usage{
validityLocus(object)
}
\arguments{
\item{object}{locus object.}
}
\value{
boolean.
}
\description{
This function tests the validity of the locus class objects.
}
