% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{validityNicheModel}
\alias{validityNicheModel}
\title{Validity function for nicheModel object}
\usage{
validityNicheModel(object)
}
\arguments{
\item{object}{nicheModel object.}
}
\description{
Validity control for the nicheModel objects.
}
