% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{validitysampledCells}
\alias{validitysampledCells}
\title{Validity function for sampledCells objects.}
\usage{
validitysampledCells(object)
}
\arguments{
\item{object}{The sampledCells object.}
}
\value{
Boolean.
}
\description{
Validity verification for sampledCells objects. It checks that the samples, the sample times and the coordinates are the same length.
}
