% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{validitysamplePoints}
\alias{validitysamplePoints}
\title{Validity function for samplePoints objects.}
\usage{
validitysamplePoints(object)
}
\arguments{
\item{object}{The samplePoints object.}
}
\value{
Boolean.
}
\description{
Validity verification for samplePoints objects. It checks that the samples, the sample times and the coordinates are the same length.
}
