% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{validitysocioecoGeoDataModel}
\alias{validitysocioecoGeoDataModel}
\title{Validity function for objects of the socioecoGeoDataModel class.}
\usage{
validitysocioecoGeoDataModel(object)
}
\arguments{
\item{object}{The evaluated socioecoGeoDataModel object.}
}
\value{
Boolean.
}
\description{
This function validates that the R and K models are \code{nicheModel} objects and that the geo-migration model is a \code{geoMigrationModel} object.
}
