% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{validitysocioecoGroupData}
\alias{validitysocioecoGroupData}
\title{Validity function for socioecoGroupData}
\usage{
validitysocioecoGroupData(object)
}
\arguments{
\item{object}{socioecoGroupData object.}
}
\description{
Validates that the number of cells and the number of categories are equal in socioecoGroupData object.
}
