% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{validitysocioecoMigrationModel}
\alias{validitysocioecoMigrationModel}
\title{Validity for socioecoMigrationModel objects.}
\usage{
validitysocioecoMigrationModel(object)
}
\arguments{
\item{object}{socioecoMigrationModel object.}
}
\value{
boolean.
}
\description{
Validates socioecoMigrationModel objects.
Checks if the length of pMig is different from length of varMig, if the migrationShape is one of the socioecoMigrationShapes elements, and if the names of pMig list are equal to the names of varMig.
}
