% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{valuesA}
\alias{valuesA}
\alias{valuesA,geoEnvData}
\title{values A.}
\usage{
\S4method{valuesA}{geoEnvData}(object)

\S4method{valuesA}{socioecoGroupData}(object)
}
\description{
values A.

valuesA method for socioecoGroupData.
}
