% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{valuesA}
\alias{valuesA}
\title{Values of the cells in a raster.}
\usage{
valuesA(object)
}
\arguments{
\item{object}{raster object.}
}
\value{
vector containing the values of the cells in a raster layer.
}
\description{
This function returns the value of the cells in a raster layer.
}
