% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{variable.names}
\alias{variable.names}
\alias{variable.names,socioecoGroupData}
\alias{variable.names,socioecoGeoData}
\alias{variable.names,socioecoMigrationModel}
\alias{variable.names,nicheModel}
\title{Variable.names method for socioecoGroupData.}
\usage{
\S4method{variable.names}{socioecoGroupData}(object)

\S4method{variable.names}{socioecoGroupsData}(object)

\S4method{variable.names}{socioecoGeoData}(object)

\S4method{variable.names}{socioecoMigrationModel}(object)

\S4method{variable.names}{geoMigrationModel}(object)

\S4method{variable.names}{nicheModel}(object)
}
\description{
Variable.names method for socioecoGroupData.

Variable.names method for socioecoGroupsData.

Variable.names method for socioecoGeoData.

variable.names method for socioecoMigrationModel.

variable.names method for geoMigrationModel.

variable.names method for nicheModel.
}
