% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{xyA}
\alias{xyA}
\title{Coordinate difference from each cell}
\usage{
xyA(object)
}
\arguments{
\item{object}{raster object.}
}
\value{
data.frame of the coordinate distance from each non-NA cell to every other.
}
\description{
Gets the coordinate distance from each non-NA cell to each other in the raster.
}
