% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{Acells}
\alias{Acells}
\alias{Acells,socioecoGroupData}
\alias{Acells,socioecoGroupsData}
\title{Acells method for socioecoGroupData.}
\usage{
\S4method{Acells}{socioecoGroupData}(object)

\S4method{Acells}{socioecoGroupsData}(object)
}
\description{
Acells method for socioecoGroupData.

Acells method for socioecoGroupsData.
}
