% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{coalescent_2_newick}
\alias{coalescent_2_newick}
\alias{coalescent_2_newick,coalSim}
\title{coalescent_2_newick method for coalSim objects.}
\usage{
\S4method{coalescent_2_newick}{coalSim}(object)
}
\description{
coalescent_2_newick method for coalSim objects.
}
