% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{locus}
\alias{locus}
\title{Create a locus object}
\usage{
locus(locusName = NULL, mValues = NULL)
}
\arguments{
\item{locusName}{Character. The name of the marker.}

\item{mValues}{Numeric. The value(s) of the marker alleles.}
}
\value{
a locus object.
}
\description{
This function creates a locus object.
}
