% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\name{plot.coalSim}
\alias{plot.coalSim}
\alias{plot}
\title{Plot method for coalSim objects}
\usage{
\method{plot}{coalSim}(object)
}
\arguments{
\item{object}{coalSim. The coalescent object.}
}
\value{
plot of the tree representing the coalescent.
}
\description{
Plot method for coalSim objects.
}
