% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,geoEnvData}
\alias{show,socioecoGroupData}
\alias{show,socioecoGroupsData}
\alias{show,socioecoGeoData}
\alias{show,locus}
\alias{show,genotype}
\alias{show,haplotype}
\alias{show,samplePoints}
\alias{show,genetSample}
\alias{show,mutationModel}
\alias{show,genetSet}
\alias{show,socioecoGeoDataHistory}
\alias{show,nicheModel}
\alias{show,geoMigrationModel}
\alias{show,socioecoGeoDataModel}
\alias{show,coalSim}
\title{Show method for geoEnvData.}
\usage{
\S4method{show}{geoEnvData}(object)

\S4method{show}{socioecoGroupData}(object)

\S4method{show}{socioecoGroupsData}(object)

\S4method{show}{socioecoGeoData}(object)

\S4method{show}{locus}(object)

\S4method{show}{genotype}(object)

\S4method{show}{haplotype}(object)

\S4method{show}{samplePoints}(object)

\S4method{show}{genetSample}(object)

\S4method{show}{mutationModel}(object)

\S4method{show}{genetSet}(object)

\S4method{show}{socioecoGeoDataHistory}(object)

\S4method{show}{nicheModel}(object)

\S4method{show}{geoMigrationModel}(object)

\S4method{show}{socioecoGeoDataModel}(object)

\S4method{show}{coalSim}(object)
}
\description{
Show method for geoEnvData.

Show method for socioecoGroupData.

show method for socioecoGroupsData.

Show method for socioecoGeoData.

show method for locus objects.

Show method for genotype objects.

Show method for haplotype objects.

show method for samplePoints objects.

show method for genetSample objects.

show method for mutationModel

show method for genetSet objects

show method for socioecoGeoDataHistory objects.

show method for nicheModel objects.

show method for geoMigrationModel objects.

show method for socioecoGeoDataModel objects.

show method for coalSim objects.
}
