% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPOP.R
\docType{methods}
\name{simulMultiCoal}
\alias{simulMultiCoal}
\alias{simulMultiCoal,ecoGenetSet,logical,numeric}
\alias{simulMultiCoal,ecoGenetSet,logical,numeric,missing}
\alias{simulMultiCoal,ecoGenetSet,logical,numeric,missing,ANY}
\title{simulMultiCoal method for ecoGenetData objects.}
\usage{
\S4method{simulMultiCoal}{ecoGenetSet,logical,numeric,logical,numeric}(
  ecoGenetSet,
  printCoal,
  iteration,
  multiCore = FALSE,
  cores = NULL
)

\S4method{simulMultiCoal}{ecoGenetSet,logical,numeric,logical,missing}(
  ecoGenetSet,
  printCoal,
  iteration,
  multiCore = FALSE,
  cores = NULL
)

\S4method{simulMultiCoal}{ecoGenetSet,logical,numeric,missing,ANY}(
  ecoGenetSet,
  printCoal,
  iteration,
  multiCore = FALSE,
  cores = NULL
)
}
\description{
simulMultiCoal method for ecoGenetData objects.

simulMultiCoal method for ecoGenetData objects.

simulMultiCoal method for ecoGenetData objects.
}
